clean up layout to use src directory
