+++ package.json
@@ -3,7 +3,7 @@
   "scripts": {
     "lint": "eslint ."
   },
-  "main": "./index.js",
+  "main": "./src/index.js",
   "version": "1.0.6",
   "author": "Andy Richardson",
   "description": "Used to check for, download and install updates to Atom Editor in linux based distributions.",
