+++ src/lib/index.ts
@@ -1,0 +1,4 @@
+import * as Notifier from './notifier';
+export const Notifer = Notifier;
+export * from './package';
+export * from './updater';
