+++ src/typings/enum.ts
