+++ src/typings/interface.d.ts
