+++ tsconfig.json
@@ -8,5 +8,9 @@
     "preserveConstEnums": true,
     "strict": true,
     "target": "es2015"
-  }
+  },
+  "include": [
+    "./src/*.ts",
+    "./src/**/*.ts"
+  ]
 }
